Refactor utilities and tests
